remove victory dep and rename components
