+++ frontend/package.json
@@ -21,8 +21,7 @@
     "react-redux": "8.0.4",
     "react-router-dom": "6.4.2",
     "react-scripts": "5.0.1",
-    "react-select": "5.5.4",
-    "victory": "^36.6.8"
+    "react-select": "5.5.4"
   },
   "devDependencies": {
     "axe-core": "4.4.3",
