+++ frontend/src/components/PortfolioFundingByBudgetStatus/PortfolioFundingByBudgetStatus.jsx
@@ -1,12 +1,13 @@
 import { getCurrentFiscalYear } from "./util";
 import { useSelector } from "react-redux";
 import CurrencyFormat from "react-currency-format";
-import { VictoryPie, Slice, VictoryLabel, VictoryTooltip } from "victory";
 
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { library } from "@fortawesome/fontawesome-svg-core";
 import { faSquare } from "@fortawesome/free-solid-svg-icons";
-import { useState } from "react";
+import React, { useState } from "react";
+
+import { CustomPie } from "./CustomPie";
 
 library.add(faSquare);
 
@@ -23,55 +24,43 @@ const colors = [
     "#429195",
 ];
 
-const CustomSlice = (props) => {
-    const [scale, setScale] = useState(1);
+const styles = {
+    root: {
+        // fontFamily: "consolas, sans-serif",
+        // textAlign: "center",
+        // position: "relative",
+        width: 150,
+        height: 150,
+        //display: "inline-block",
+    },
+    cardBody: {
+        //display: "inline-block",
+    },
+};
 
-    // modified transformation from here
-    // https://github.com/FormidableLabs/victory/blob/844109cfe4e40b23a4dcb565e551a5a98015d0c0/packages/victory-pie/src/slice.js#L74
-    const transform = `translate(${props.origin.x}, ${props.origin.y}) scale(${scale})`;
+const CustomLayerComponent = (myProps) => (layerProps) => {
+    const { centerX, centerY } = layerProps;
+
+    console.log(`myProps=${{ myProps }}`);
+    console.log(layerProps);
 
     return (
-        <Slice
-            {...props}
-            style={{ ...props.style }}
-            // events={{
-            //     onMouseOver: (e) => {
-            //         if (props.events.onMouseOver) {
-            //             // props.events.onMouseOver(e);
-            //         }
-            //         // setScale((c) => c * 1.2);
-            //     },
-            //     onMouseOut: (e) => {
-            //         if (props.events.onMouseOut) {
-            //             // props.events.onMouseOut(e);
-            //         }
-            //         // setScale(1);
-            //     },
-            // }}
-            events={{
-                onMouseOver: (e) => {
-                    const t = e.target.style.fill;
-                    console.log({ t });
-                    props.setPercent("100%");
-                    // if (props.events.onMouseOver) {
-                    //     props.events.onMouseOver(e);
-                    // }
-                    setScale((c) => c * 1.2);
-                },
-                onMouseOut: (e) => {
-                    props.setPercent("0%");
-                    // if (props.events.onMouseOut) {
-                    //     props.events.onMouseOut(e);
-                    // }
-                    setScale(1);
-                },
+        <text
+            x={centerX}
+            y={centerY}
+            textAnchor="middle"
+            dominantBaseline="central"
+            style={{
+                fontSize: "20px",
+                fontWeight: "600",
             }}
-            transform={transform}
-        />
+        >
+            {myProps}
+        </text>
     );
 };
 
-const PortfolioFundingByBudgetStatus = () => {
+const PortfolioFundingByBudgetStatus = (props) => {
     const today = new Date();
     const portfolioFunding = useSelector((state) => state.portfolioFundingSummary.portfolioFunding);
     const portfolioFundingChart = useSelector((state) => state.portfolioFundingSummary.portfolioFundingChart);
@@ -85,55 +74,6 @@ const PortfolioFundingByBudgetStatus = () => {
                         FY {getCurrentFiscalYear(today)} Budget Status
                     </h3>
                 </div>
-            </div>
-            <div className="usa-card__media padding-1">
-                <VictoryPie
-                    dataComponent={<CustomSlice setPercent={setPercent} />}
-                    className="usa-card__img margin-top-2"
-                    width={700}
-                    height={700}
-                    aria-label="Portfolio Donut Chart With Budget Status Percentages"
-                    data={portfolioFundingChart}
-                    innerRadius={200}
-                    labelRadius={100}
-                    labels={() => null}
-                    style={{
-                        data: {
-                            fill: ({ datum }) => datum.fill,
-                        },
-                    }}
-                    labelComponent={
-                        <VictoryLabel
-                            textAnchor="middle"
-                            verticalAnchor="middle"
-                            style={{ fontSize: 40 }}
-                            x={350}
-                            y={350}
-                            text={percent}
-                        />
-                    }
-                />
-                {/*<div className="usa-card__media padding-1">*/}
-                {/*    <VictoryPie*/}
-                {/*        dataComponent={<CustomSlice />}*/}
-                {/*        labelComponent={*/}
-                {/*            <VictoryTooltip*/}
-                {/*                style={{ fontSize: 30, fill: "#FFFFFF" }}*/}
-                {/*                flyoutStyle={{ fill: "#000000" }}*/}
-                {/*                constrainToVisibleArea*/}
-                {/*            />*/}
-                {/*        }*/}
-                {/*        className="usa-card__img margin-top-2"*/}
-                {/*        aria-label="Portfolio Pie Chart With Funding Status Percentages"*/}
-                {/*        data={portfolioFundingChart}*/}
-                {/*        labels={[]}*/}
-                {/*        style={{*/}
-                {/*            data: {*/}
-                {/*                fill: ({ datum }) => datum.fill,*/}
-                {/*            },*/}
-                {/*        }}*/}
-                {/*    />*/}
-                {/*</div>*/}
             </div>
             <div className="usa-card__body padding-1">
                 <div className="grid-container padding-top-0 padding-1 font-sans-3xs">
@@ -203,6 +143,17 @@ const PortfolioFundingByBudgetStatus = () => {
                                     {value}
                                 </div>
                             )}
+                        />
+                    </div>
+                </div>
+            </div>
+            <div className="usa-card__media usa-card__media--inset">
+                <div className="usa-card__img">
+                    <div style={styles.root}>
+                        <CustomPie
+                            data={portfolioFundingChart}
+                            CustomLayerComponent={CustomLayerComponent(percent)}
+                            setPercent={setPercent}
                         />
                     </div>
                 </div>
