--- frontend/src/components/PortfolioFundingByBudgetStatusNivio/util.js
@@ -1,10 +1,0 @@
-export const getCurrentFiscalYear = (today) => {
-    const currentMonth = today.getMonth();
-
-    let fiscalYear;
-    currentMonth < 9
-        ? (fiscalYear = today.getFullYear().toString())
-        : (fiscalYear = (today.getFullYear() + 1).toString());
-
-    return fiscalYear;
-};
