--- frontend/src/components/PortfolioFundingByBudgetStatusNivio/util.test.js
@@ -1,9 +1,0 @@
-import { getCurrentFiscalYear } from "./util";
-
-test("current federal fiscal year is calculated correctly", async () => {
-    const lastDay = new Date("September 30, 2022");
-    const firstDay = new Date("October 1, 2022");
-
-    expect(getCurrentFiscalYear(lastDay)).toEqual("2022");
-    expect(getCurrentFiscalYear(firstDay)).toEqual("2023");
-});
