+++ frontend/src/pages/portfolios/detail/PortfolioDetail.jsx
@@ -28,27 +28,22 @@ const PortfolioDetail = () => {
     }, [dispatch, portfolioId]);
 
     return (
-        <main>
-            <div className="grid-container">
-                <div className="grid-row">
-                    <h1>{portfolio.name}</h1>
-                </div>
-                <div className="grid-row">
-                    <h2>Portfolio description</h2>
-                    {portfolio.description}
-                </div>
-                <div className="grid-row">
-                    <PortfolioFundingSummary portfolioId={portfolioId} />
-                </div>
-                <div className="grid-row">
-                    <h2>CANs</h2>
-                    <ul className="usa-list">
-                        {portfolio.internal_can?.map((can) => (
-                            <CanList key={can.id} id={can.id} name={can.number} />
-                        ))}
-                    </ul>
-                </div>
-            </div>
+        <main className="margin-left-5 margin-right-5">
+            <section>
+                <h1 className="font-sans-2xl">{portfolio.name}</h1>
+                {portfolio.description}
+            </section>
+            <section>
+                <PortfolioFundingSummary portfolioId={portfolioId} />
+            </section>
+            <section>
+                <h2>CANs</h2>
+                <ul className="usa-list">
+                    {portfolio.internal_can?.map((can) => (
+                        <CanList key={can.id} id={can.id} name={can.number} />
+                    ))}
+                </ul>
+            </section>
         </main>
     );
 };
